refactor: Update Viewer folder structure
