+++ packages/insight-viewer/src/Viewer/index.ts
@@ -1,3 +1,2 @@
 export { DICOMImageViewer } from './DICOMImageViewer'
 export { WebImageViewer } from './WebImageViewer'
-export { default as ViewContext } from './Context'
